+++ lib/active_date_range.rb
@@ -15,6 +15,10 @@ require "active_date_range/active_model_type"
 require "active_date_range/validators" if defined?(ActiveModel)
 
 module ActiveDateRange
+  def self.deprecator
+    @deprecator ||= ActiveSupport::Deprecation.new("0.6.0", "ActiveDateRange")
+  end
+
   class Error < StandardError; end
   class InvalidDateRange < Error; end
   class InvalidAddition < Error; end