+++ test/active_date_range/active_model_type_test.rb
@@ -22,27 +22,43 @@ class ActiveDateRangeAttributesTest < ActiveSupport::TestCase
     assert_raises(ActiveDateRange::InvalidDateRangeFormat) { report.period }
   end
 
-  class SafeTestReport
+  class SafeParseTestReport
     include ActiveModel::Attributes
 
-    attribute :period, :date_range, safe: true
+    attribute :period, :date_range, safe_parse: true
   end
 
-  def test_safe_date_range_conversion
-    report = SafeTestReport.new
+  def test_safe_parse_date_range_conversion
+    report = SafeParseTestReport.new
     report.period = "this_month"
     assert_equal ActiveDateRange::DateRange.this_month, report.period
   end
 
-  def test_safe_date_range_conversion_returns_nil
-    report = SafeTestReport.new
+  def test_safe_parse_date_range_conversion_returns_nil
+    report = SafeParseTestReport.new
     report.period = "unknown"
     assert_nil report.period
   end
 
-  def test_safe_date_range_conversion_returns_nil_for_reversed_range
-    report = SafeTestReport.new
+  def test_safe_parse_date_range_conversion_returns_nil_for_reversed_range
+    report = SafeParseTestReport.new
     report.period = "202503..202501"
+    assert_nil report.period
+  end
+
+  def test_deprecated_safe_option_still_works
+    report_class = nil
+
+    assert_deprecated("safe_parse", ActiveDateRange.deprecator) do
+      report_class = Class.new do
+        include ActiveModel::Attributes
+
+        attribute :period, :date_range, safe: true
+      end
+    end
+
+    report = report_class.new
+    report.period = "unknown"
     assert_nil report.period
   end
 end